Update styles for description list example
For tablet and up, set a width for the counter number

This ensures it will line up with the task list content underneath.
For smaller screens, where space is at a premium, don’t line up this
content - instead ensure it is all left-aligned.

Increase spacing to the right of the task section number.